MR fixes

See merge request dart-packages/layer-sdk!366

+++ lib/presentation_layer/cubits/loyalty/loyalty_redemption/loyalty_redemption_cubit.dart
@@ -161,7 +161,7 @@ class LoyaltyRedemptionCubit extends Cubit<LoyaltyRedemptionState> {
             LoyaltyRedemptionAction.rate,
           ),
           loyaltyPoints: state.loyaltyPoints.copyWith(
-            rate: rate.rate.toDouble(),
+            rate: rate.rate,
           ),
         ),
       );
